requirements/base.txt
# Base requirements. These are the only ones that are required to run the app.

# Core for web.
django==4.2.7

# For environment variables and configuration.
django-environ==0.11.2

# For PostgreSQL.
# psycopg2-binary==2.9.7
psycopg[binary,pool]==3.1.12


# Dependencies for image field
Pillow==10.1.0

# For filters
django-filter==23.3


# [drf]-[BEGIN]
djangorestframework==3.14.0

# Docs
drf-yasg==1.21.7
drf-spectacular==0.27.0
drf-spectacular-sidecar==2023.12.1
# [drf]-[END]

# Countries and Cities
# googlemaps~=4.10.0 (if we would use API Place from Google)
django-cities-light==3.10.0


# Authentication
django-allauth==0.58.2
dj_rest_auth==5.0.1


# phone number's validating
django-phonenumber-field==7.2.0
phonenumbers==8.13.24

celery[redis]==5.3.5
flower==2.0.1

# Sender sms for validation phonenumbers
twilio==8.10.2
vonage==3.11.0

alphasms-client==0.1.3


django-cors-headers==4.3.1

# For pytest.
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-mock==3.12.0

gunicorn==21.2.0
pika==1.3.2

djangorestframework-simplejwt==5.3.1
pandas==2.1.4
openpyxl==3.1.2

setuptools

# work with data
pandas==2.1.4
openpyxl==3.1.2

# google auth
google-auth==2.25.2

pymemcache

google-cloud-storage==2.14.0

# AWS utils
boto3
botocore